matrix

+++ 0x07-python-test_driven_development/2-matrix_divided.py
@@ -29,4 +29,4 @@ def matrix_divided(matrix, div):
     if not isinstance(div, (int, float)):
         raise TypeError("div must be a number")
     
-    return [[("{:.2f}".format(num/div)) for num in row] for row in matrix]
+    return [[float("{:.2f}".format(num/div)) for num in row] for row in matrix]
